requirement.txt
astroid==2.3.1
colorama==0.4.1
Django==2.2.6
django-appconf==1.0.3
django-imagekit==4.0.2
django-resized==0.3.9
django-tinymce==2.8.0
django-tinymce4-lite==1.7.5
docopt==0.6.2
html5lib==1.0.1
image==1.5.27
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.2
mccabe==0.6.1
mysql-connector==2.2.9
mysqlclient==1.4.4
num2words==0.5.10
numpy==1.17.4
pandas==0.25.3
pilkit==2.0
Pillow==6.2.0
pylint==2.4.2
PyPDF2==1.26.0
python-barcode==0.10.0
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.32
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.6
virtualenvwrapper-win==1.2.5
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
